requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
requests==2.18.4
todoist-python==7.0.17
urllib3==1.22
youtrack-rest-api==0.0.1
